removed RENAME_FILE process, and made it a utility func in helpers.py
